Add IStream to abstract abstract Stream class and add support in IBuffer as IStream.  Refactored Lzma hierarchy and have load test working (commented out).

+++ src/FortitudeCommon/DataStructures/Memory/UnmanagedMemory/UnmanagedByteArray.cs
@@ -16,7 +16,7 @@ public unsafe class UnmanagedByteArray : IByteArray
 
     private VirtualMemoryByteArrayEnumerator? reusableEnumerator;
 
-    public UnmanagedByteArray(IVirtualMemoryAddressRange mappedViewRegion, long arrayOffset, int length)
+    public UnmanagedByteArray(IVirtualMemoryAddressRange mappedViewRegion, long arrayOffset, long length)
     {
         if (mappedViewRegion.SizeBytes < arrayOffset + length)
             throw new Exception("Memory mapped file view size does not match expected file position and/or size");
